refactor: reorganize editor type imports for better structure
