+++ src/types/index.d.ts
@@ -18,5 +18,4 @@ export * from './participant';
 export * from './search';
 export * from './share';
 export * from './soap';
-export * from './state';
 export * from './utils';
