+++ src/views/search/list/conversation/search-conversation-messages-list.tsx
@@ -13,7 +13,8 @@ import { useNavigate } from 'react-router-dom';
 
 import { API_REQUEST_STATUS } from 'constants/index';
 import { useMultipleSelection } from 'hooks/use-multiple-selection';
-import { IncompleteMessage, SearchRequestStatus } from 'types/index.d';
+import { IncompleteMessage } from 'types/messages';
+import { SearchRequestStatus } from 'types/search';
 import { MessageListItem } from 'views/app/folder-panel/messages/message-list-item';
 
 type SearchConversationMessagesListProps = {
